requirements.txt
aniso8601==9.0.1
Cerberus==1.3.2
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
importlib==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.0
Werkzeug==1.0.1
